requirements.txt
Flask
flask_sqlalchemy
requests
pandas
numpy
ta
python-dotenv
openai
reportlab
google-generativeai